# web6/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FundedFrens API — Solana meme coin prop trading platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/users/profile:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get authenticated user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update authenticated user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, auth_user_id, email, role, telegram_status, created_at, updated_at]
      properties:
        id:
          type: integer
        auth_user_id:
          type: string
        email:
          type: string
        username:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        telegram_status:
          type: string
          enum: [linked, not_linked]
        telegram_username:
          type: ["string", "null"]
        wallet_address:
          type: ["string", "null"]
        referral_code:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    ProfileUpdate:
      type: object
      properties:
        username:
          type: string
        wallet_address:
          type: string
...